Refactor e2e tests
